requirements.txt
fastapi
uvicorn
python-dotenv
openai
numpy
opencv-python
python-multipart